Put JdbcBaseDal into -lib-db
